pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "og"
version = "0.1.0"
description = "OpenClaw Python PoC — CLI agent with Pi agent patterns"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "click>=8.0",
    "aiofiles>=24.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
og = "og.__main__:main"

[tool.ruff]
target-version = "py311"
line-length = 100
